refactoring erros
