+++ src/main/business/stockMovementBusiness.js
@@ -1,4 +1,5 @@
-import CustomError from "../util/CustomError.js"
+import { EntityNotFound } from "../erros/EntityNotFoundError.js"
+import { StockAdjustmentError } from "../erros/StockAdjustmentError.js"
 import db from "../database/db.js"
 import { adjustStock } from "../util/stockMovementHelper.js"
 import { Op } from "sequelize"
@@ -10,7 +11,7 @@ export const create = async (event, stockMovementData) => {
         const product = await db.Product.findByPk(stockMovementData.productId)
 
         if (!product) {
-            throw new CustomError('Produto associado não encontrado.')
+            throw new EntityNotFound('Produto associado não encontrado.')
         }
 
         const { type, quantity } = stockMovementData
@@ -18,7 +19,7 @@ export const create = async (event, stockMovementData) => {
         const newStock = adjustStock(product.currentStock, type, quantity)
 
         if (newStock < 0) {
-            throw new CustomError('O movimento causaria um estoque negativo.')
+            throw new StockAdjustmentError('O movimento causaria um estoque negativo.')
         }
 
         await product.update({ currentStock: newStock }, { transaction })
@@ -39,7 +40,7 @@ export const findById = async (event, id) => {
     const stockMovementFound = await db.StockMovement.findByPk(id)
 
     if (!stockMovementFound) {
-        throw new CustomError("Não existe Movimento de Estoque com esse id")
+        throw new EntityNotFound("Não existe Movimento de Estoque com esse id")
     }
 
     return stockMovementFound.dataValues
@@ -62,13 +63,13 @@ export const update = async (event, stockMovementData) => {
         const existingStockMovement = await db.StockMovement.findByPk(id)
 
         if (!existingStockMovement) {
-            throw new CustomError("Não existe Movimento de Estoque com esse id")
+            throw new EntityNotFound("Não existe Movimento de Estoque com esse id")
         }
 
         const product = await db.Product.findByPk(existingStockMovement.productId)
 
         if (!product) {
-            throw new CustomError('Produto associado não encontrado.')
+            throw new EntityNotFound('Produto associado não encontrado.')
         }
 
         const { type: esmType, quantity: esmQty } = existingStockMovement
@@ -78,7 +79,7 @@ export const update = async (event, stockMovementData) => {
         const newStock = adjustStock(adjustedStock, smdType, smdQty)
 
         if (newStock < 0) {
-            throw new CustomError('O movimento causaria um estoque negativo.')
+            throw new StockAdjustmentError('O movimento causaria um estoque negativo.')
         }
 
         await product.update({ currentStock: newStock }, { transaction })
@@ -98,21 +99,17 @@ export const destroy = async (event, id) => {
         const stockMovementFound = await db.StockMovement.findByPk(id)
 
         if (!stockMovementFound) {
-            throw new CustomError("Não existe Movimento de Estoque com esse id")
+            throw new EntityNotFound("Não existe Movimento de Estoque com esse id")
         }
 
         const product = await db.Product.findByPk(stockMovementFound.productId)
 
         if (!product) {
-            throw new CustomError('Produto associado não encontrado.')
+            throw new EntityNotFound('Produto associado não encontrado.')
         }
 
         const { type, quantity } = stockMovementFound
         const newStock = adjustStock(product.currentStock, type, -quantity)
-
-        if (newStock < 0) {
-            throw new CustomError('O movimento causaria um estoque negativo ao ser deletado.')
-        }
 
         await product.update({ currentStock: newStock }, { transaction })
         await stockMovementFound.destroy({ transaction })
