+++ src/shared/errorTypes.js
@@ -2,5 +2,6 @@ export const ErrorTypes = {
     ENTITY_ALREADY_EXISTS: 'ENTITY_ALREADY_EXISTS',
     ENTITY_NOT_FOUND:'ENTITY_NOT_FOUND',
     SAVE_IMAGE:'SAVE_IMAGE',
+    STOCK_ADJUSTMENT:'STOCK_ADJUSTMENT',
     INTERNAL: 'INTERNAL_ERROR'
 }
